refactor(video): restructure video module into package
- Move video.py functionality into video/ package
- Split video reader and writer into separate modules
- Add video typing definitions
- Update video decoder imports
- Simplify multi_gpu_loader script

BREAKING CHANGE: video module API has been restructured from single file to package
