+++ projects/owa-core/owa/core/io/video/typing.py
@@ -1,0 +1,5 @@
+from fractions import Fraction
+from typing import Literal, Union
+
+SECOND_TYPE = Union[float, Fraction]
+PTSUnit = Literal["pts", "sec"]
